Prefer GitHub Container Registry over DockerHub
We want to use GitHub so that everything is hosted under one roof, and we don't have a bunch of other accounts to manage. We couldn't use GHCR previously because it had such strict permissions as to be unusable within GitHub Actions. But anonymous read access to images is now available, allowing GafferHQ/Gaffer's CI workflows to pick up the images easily.