build.py : Include `GAFFER_BUILD_ENVIRONMENT` in dependencies package name
This also means that we're now using `--package` to explicitly control the name of the file produced by GafferHQ/dependencies, rather than guessing what it is going to be. The general principle is that `GafferHQ/gaffer` and `GafferHQ/dependencies` are non-opinionated about the build environment or package name, allowing both to be controlled by some external environment (`GafferHQ/build`, Gaffer's CI, personal build configs, etc).